Fix GetTaskExitStatus success criteria
Proxmox had changed its API:
On March 2024, it changed an error `can't deactivate LV` when cloning from template in parallel. It got downgraded to a warning, which means that the clone tasks will continue to be completed successfully. Therefore, we need to mark this as a successful task.

See the following:
* https://lists.proxmox.com/pipermail/pve-devel/2024-March/062115.html
* https://github.com/Telmate/proxmox-api-go/commit/e79247d189085f2e20426a85a9748701b1816129